requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.4
python-dotenv==1.0.1
groq==0.4.2
qdrant-client==1.8.0
langchain-qdrant==0.1.1
langchain-ollama>=0.1.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
streamlit>=1.30.0
